+++ src/main/resources/templates/index.html
@@ -28,7 +28,7 @@
 				<div class="mdl-layout-spacer"></div>
 				<!-- Tabs -->
 				<nav class="mdl-navigation">
-					<a class="mdl-navigation__link" href="/">Issues</a>
+					<a class="mdl-navigation__link" href="/issue/">Issues</a>
 					<a class="mdl-navigation__link" href="/milestone/list.html">Milestones</a>
 					<a class="mdl-navigation__link" href="/user/login.html">login</a>
 					<a class="mdl-navigation__link" href="/user/join.html">join</a>
